+++ example_runner.py
@@ -67,6 +67,7 @@ print('material examples')
 import example.material.stones
 import example.material.uneven_blocks
 import example.material.uneven_centered_blocks
+import example.material.stucco_brick_blocks
 
 print('minibase examples')
 import example.minibase.circle