+++ src/cqterrain/material/__init__.py
@@ -1,4 +1,6 @@
-from .stone import make_stones
+from .make_stones import make_stones
 from .center_blocks import center_blocks
 from .uneven_blocks import uneven_blocks
 from .uneven_centered_blocks import uneven_centered_blocks
+from .stacked_wave_form_map import stacked_wave_form_map
+from .stucco_brick_blocks import stucco_brick_blocks