+++ changes.md
@@ -1,5 +1,11 @@
 ## Main wip
 
+## 3.2.0
+* Added material stacked_wave_from_map
+* Added material stucco_brick_blocks
+* Added examples for stucco_brick_blocks
+* Updated metarial documentation
+
 ## 3.1.0
 * Added build_plate for CrystalWall
 * Added build_assembly for CrystalWall